requirements.txt
antlr4-python3-runtime==4.8
anyio==3.6.1
arrow==1.2.2
asgiref==3.5.2
attrs==21.4.0
binaryornot==0.4.4
black==22.3.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cookiecutter==1.7.3
fastapi==0.78.0
h11==0.13.0
httptools==0.4.0
hydra-core==1.1.2
idna==3.3
importlib-resources==5.2.3
iniconfig==1.1.1
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.1.0
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-dataclass==8.5.8
mypy==0.960
mypy-extensions==0.4.3
numpy==1.22.3
omegaconf==2.1.2
packaging==21.3
pandas==1.4.2
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
poyo==0.5.0
py==1.11.0
pydantic==1.9.1
pyparsing==3.0.8
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==6.1.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
text-unidecode==1.3
threadpoolctl==3.1.0
tomli==2.0.1
types-PyYAML==6.0.7
typing-inspect==0.7.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
zipp==3.8.0
